extracting common methods to concern
